requirements.txt
streamlit
google-generativeai
gtts
Pillow
pokemontcgsdk
pydub